+++ dist.sh
@@ -1,1 +1,1 @@
-pyinstaller pcsiGUI.py --hidden-import 'packaging.requirements' --hidden-import 'pkg_resources.py2_warn'
+pyinstaller pcsiGUI.spec