+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Issue #8: Enable hardware rts/cts handshaking and ignore data before the first
+ start flag.
+- Issue #9: Explicitly import tkinter to help pyinstaller find it
+- Issue #10: Display the percentage of packets received instead of the percentage of pixels.
+
+
 ## [0.0.0] - 2020-06-28
 
 ### Added